[reaper] add S3 bucket and log IDs for soft and hard reaping there permanently

+++ reaper/src/index.ts
@@ -1,6 +1,11 @@
 import type { Response } from 'node-fetch';
 import fetch from 'node-fetch';
-import { getBatchSize, getMediaApiHostname, getMediaApiKey } from './util';
+import {
+	getBatchSize,
+	getMediaApiHostname,
+	getMediaApiKey,
+	storeRecord,
+} from './util';
 
 type HardDeletedStatuses = Record<
 	string, // media ID
@@ -32,6 +37,8 @@ export const handler = async () => {
 		},
 	};
 
+	const recordDate = new Date();
+
 	const IDsToSoftDelete = await fetch(
 		`${mediaApiImagesBaseUrl}/nextIdsToBeSoftReaped?size=${batchSize}`,
 		fetchOptions,
@@ -40,47 +47,57 @@ export const handler = async () => {
 		message: `${IDsToSoftDelete.length} images for soft deletion.`,
 		IDsToSoftDelete,
 	});
-
-	const softDeleteResponse = await fetch(
-		`${mediaApiImagesBaseUrl}/batchSoftDelete`,
-		{
-			...fetchOptions,
-			method: 'DELETE',
-			body: JSON.stringify(IDsToSoftDelete),
-		},
-	);
-	if (!softDeleteResponse.ok) {
-		throw new Error(
-			`Soft delete failed: ${softDeleteResponse.status} ${softDeleteResponse.statusText}`,
+	if (IDsToSoftDelete.length > 0) {
+		const softDeleteResponse = await fetch(
+			`${mediaApiImagesBaseUrl}/batchSoftDelete`,
+			{
+				...fetchOptions,
+				method: 'DELETE',
+				body: JSON.stringify(IDsToSoftDelete),
+			},
 		);
+		if (!softDeleteResponse.ok) {
+			throw new Error(
+				`Soft delete failed: ${softDeleteResponse.status} ${softDeleteResponse.statusText}`,
+			);
+		}
+		console.log({ message: 'Soft delete succeeded.', IDsToSoftDelete });
+		// TODO after delay consider ES check of all the IDs to ensure they're soft deleted
+		await storeRecord(recordDate, { softDeleted: IDsToSoftDelete });
 	}
-	console.log({ message: 'Soft delete succeeded.', IDsToSoftDelete });
-	// TODO after delay consider ES check of all the IDs to ensure they're soft deleted
-	// TODO log IDs to permanent location
 
 	const IDsToHardDelete: string[] = await fetch(
 		`${mediaApiImagesBaseUrl}/nextIdsToBeHardReaped?size=${batchSize}`,
 		fetchOptions,
 	).then(parseResponseAs<string[]>());
 	console.log({
-		message: `${IDsToHardDelete.length} images for hard deletion:`,
+		message: `${IDsToHardDelete.length} images for hard deletion`,
 		IDsToHardDelete,
 	});
+	if (IDsToHardDelete.length > 0) {
+		await storeRecord(recordDate, {
+			softDeleted: IDsToSoftDelete,
+			hardDeleted: IDsToHardDelete,
+		});
 
-	const hardDeletedStatuses = await fetch(
-		`${mediaApiImagesBaseUrl}/batchHardDelete`,
-		{
-			...fetchOptions,
-			method: 'DELETE',
-			body: JSON.stringify(IDsToHardDelete),
-		},
-	).then(parseResponseAs<HardDeletedStatuses>());
-	// TODO check for false values in the response
-	console.log({
-		message: 'Hard delete succeeded.',
-		IDsToHardDelete,
-		hardDeletedStatuses,
-	});
-	// TODO after delay consider ES check of all the IDs to ensure they're hard deleted
-	// TODO log IDs to permanent location
+		const hardDeletedStatuses = await fetch(
+			`${mediaApiImagesBaseUrl}/batchHardDelete`,
+			{
+				...fetchOptions,
+				method: 'DELETE',
+				body: JSON.stringify(IDsToHardDelete),
+			},
+		).then(parseResponseAs<HardDeletedStatuses>());
+		// FIXME check for false values in the response
+		console.log({
+			message: 'Hard delete succeeded.',
+			IDsToHardDelete,
+			hardDeletedStatuses,
+		});
+		// TODO after delay consider ES check of all the IDs to ensure they're hard deleted
+		await storeRecord(recordDate, {
+			softDeleted: IDsToSoftDelete,
+			hardDeleted: hardDeletedStatuses,
+		});
+	}
 };
